third_party/WebKit/LayoutTests/http/tests/inspector/console-fetch-logging-expected.txt

CONSOLE MESSAGE: line 9: sending a GET request to resources/xhr-exists.html
CONSOLE MESSAGE: line 9: sending a GET request to resources/xhr-does-not-exist.html
CONSOLE MESSAGE: line 9: sending a POST request to resources/post-target.cgi
CONSOLE MESSAGE: line 9: sending a GET request to http://localhost:8000/inspector/resources/xhr-exists.html
CONSOLE ERROR: Fetch API cannot load http://localhost:8000/inspector/resources/xhr-exists.html. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access. If an opaque response serves your needs, set the request's mode to 'no-cors' to fetch the resource with CORS disabled.
CONSOLE MESSAGE: line 9: sending a GET request to resources/xhr-exists.html
CONSOLE MESSAGE: line 9: sending a GET request to resources/xhr-does-not-exist.html
CONSOLE MESSAGE: line 9: sending a POST request to resources/post-target.cgi
CONSOLE MESSAGE: line 9: sending a GET request to http://localhost:8000/inspector/resources/xhr-exists.html
CONSOLE ERROR: Fetch API cannot load http://localhost:8000/inspector/resources/xhr-exists.html. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access. If an opaque response serves your needs, set the request's mode to 'no-cors' to fetch the resource with CORS disabled.
Tests that fetch logging works when XMLHttpRequest Logging is Enabled and doesn't show logs when it is Disabled.

console-fetch-logging.html:4 sending a GET request to resources/xhr-exists.html
network-test.js:48 Fetch complete: GET "http://127.0.0.1:8000/inspector/resources/xhr-exists.html".makeFetch @ network-test.js:48requestHelper @ console-fetch-logging.html:11step1 @ console-fetch-logging.html:24makeRequests @ console-fetch-logging.html:18(anonymous) @ VM:1
console-fetch-logging.html:4 sending a GET request to resources/xhr-does-not-exist.html
network-test.js:48 GET http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html 404 (Not Found)makeFetch @ network-test.js:48requestHelper @ console-fetch-logging.html:11step2 @ console-fetch-logging.html:30
network-test.js:48 Fetch complete: GET "http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html".makeFetch @ network-test.js:48requestHelper @ console-fetch-logging.html:11step2 @ console-fetch-logging.html:30
console-fetch-logging.html:4 sending a POST request to resources/post-target.cgi
network-test.js:48 Fetch complete: POST "http://127.0.0.1:8000/inspector/resources/post-target.cgi".makeFetch @ network-test.js:48requestHelper @ console-fetch-logging.html:11step3 @ console-fetch-logging.html:36
console-fetch-logging.html:4 sending a GET request to http://localhost:8000/inspector/resources/xhr-exists.html
console-fetch-logging.html:1 Fetch API cannot load http://localhost:8000/inspector/resources/xhr-exists.html. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access. If an opaque response serves your needs, set the request's mode to 'no-cors' to fetch the resource with CORS disabled.
console-fetch-logging.html:4 sending a GET request to resources/xhr-exists.html
console-fetch-logging.html:4 sending a GET request to resources/xhr-does-not-exist.html
network-test.js:48 GET http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html 404 (Not Found)makeFetch @ network-test.js:48requestHelper @ console-fetch-logging.html:11step2 @ console-fetch-logging.html:30
console-fetch-logging.html:4 sending a POST request to resources/post-target.cgi
console-fetch-logging.html:4 sending a GET request to http://localhost:8000/inspector/resources/xhr-exists.html
console-fetch-logging.html:1 Fetch API cannot load http://localhost:8000/inspector/resources/xhr-exists.html. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access. If an opaque response serves your needs, set the request's mode to 'no-cors' to fetch the resource with CORS disabled.
